+++ ubiquitous_bash.sh
@@ -36,7 +36,7 @@ _ub_cksum_special_derivativeScripts_contents() {
 #export ub_setScriptChecksum_disable='true'
 ( [[ -e "$0".nck ]] || [[ "${BASH_SOURCE[0]}" != "${0}" ]] || [[ "$1" == '--profile' ]] || [[ "$1" == '--script' ]] || [[ "$1" == '--call' ]] || [[ "$1" == '--return' ]] || [[ "$1" == '--devenv' ]] || [[ "$1" == '--shell' ]] || [[ "$1" == '--bypass' ]] || [[ "$1" == '--parent' ]] || [[ "$1" == '--embed' ]] || [[ "$1" == '--compressed' ]] || [[ "$0" == "/bin/bash" ]] || [[ "$0" == "-bash" ]] || [[ "$0" == "/usr/bin/bash" ]] || [[ "$0" == "bash" ]] ) && export ub_setScriptChecksum_disable='true'
 export ub_setScriptChecksum_header='2591634041'
-export ub_setScriptChecksum_contents='3586309678'
+export ub_setScriptChecksum_contents='3342391606'
 
 # CAUTION: Symlinks may cause problems. Disable this test for such cases if necessary.
 # WARNING: Performance may be crucial here.
@@ -44433,9 +44433,9 @@ _create_ubDistBuild-rotten_install-kde() {
 	imagedev=$(cat "$scriptLocal"/imagedev)
 	
 	[[ ! -e "$scriptLib"/custom/package_kde.tar.xz ]] && [[ ! -e "$scriptLocal"/custom/package_kde.tar.xz ]] && [[ ! -e "$scriptLocal"/package_kde.tar.xz ]] && _messageFAIL
-	sudo -n cp -f "$scriptLib"/custom/package_kde.tar.xz "$globalVirtFS"/package_kde.tar.xz
-	sudo -n cp -f "$scriptLocal"/custom/package_kde.tar.xz "$globalVirtFS"/package_kde.tar.xz
-	sudo -n cp -f "$scriptLocal"/package_kde.tar.xz "$globalVirtFS"/package_kde.tar.xz
+	[[ -e "$scriptLib"/custom/package_kde.tar.xz ]] && sudo -n cp -f "$scriptLib"/custom/package_kde.tar.xz "$globalVirtFS"/package_kde.tar.xz
+	[[ -e "$scriptLocal"/custom/package_kde.tar.xz ]] && sudo -n cp -f "$scriptLocal"/custom/package_kde.tar.xz "$globalVirtFS"/package_kde.tar.xz
+	[[ -e "$scriptLocal"/package_kde.tar.xz ]] && sudo -n cp -f "$scriptLocal"/package_kde.tar.xz "$globalVirtFS"/package_kde.tar.xz
 	[[ ! -e "$globalVirtFS"/package_kde.tar.xz ]] && _messageFAIL
 	sudo -n chmod 644 "$globalVirtFS"/package_kde.tar.xz
